Move pod analyzer to utils
