+++ federation/pkg/federation-controller/util/podanalyzer/pod_helper.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package replicaset
+package podanalyzer
 
 import (
 	"fmt"
